reorganize: hooks go in hooks folder
